refactor zipkin tracer code into separate files
